move items into settings page
