+++ src/components/nav/NavBar.tsx
@@ -1,12 +1,10 @@
 import { BaseBarComponents } from "./BaseBarComponents";
-import { LanguageSelection } from "./LanguageSelection";
 
 export function BottomBar({ width }: { width: number }) {
   return (
     <nav className="bottomNav">
       <ul>
         <BaseBarComponents width={width} />
-        <LanguageSelection width={width} />
       </ul>
     </nav>
   );
@@ -17,7 +15,6 @@ export function SideBar({ width }: { width: number }) {
     <nav className="sideNav">
       <ul>
         <BaseBarComponents width={width} />
-        <LanguageSelection width={width} />
       </ul>
     </nav>
   );
