+++ src/i18n/fr/locales.json
@@ -4,7 +4,8 @@
       "home": "Accueil",
       "lines": "Lignes",
       "about": "Aide",
-      "lang": "Langue"
+      "lang": "Langue",
+      "settings": "Paramètres"
     }
   },
   "home": {
