+++ src/i18n/zh/locales.json
@@ -4,7 +4,8 @@
       "home": "主頁",
       "lines": "路線",
       "about": "關於",
-      "lang": "語言"
+      "lang": "語言",
+      "settings": "設定"
     }
   },
   "home": {
